requirements.txt
fastapi
uvicorn
pydantic
pymupdf
pandas
litellm
python-multipart
jinja2
openpyxl
python-dotenv
